[REFACTOR] Refactor notification service to use interface for abstraction
